Move test setup code to conftest file
